requirements.txt
pyrogram==2.0.106
pymongo==4.6.0
python-dotenv==1.0.0
ffmpeg-python==0.2.0